+++ custom_components/dessmonitor/const.py
@@ -128,6 +128,20 @@ SENSOR_TYPES = {
         "state_class": "measurement",
         "icon": "mdi:battery",
     },
+    "Battery Charging Current": {
+        "name": "Battery Charging Current",
+        "unit": UNITS["CURRENT"],
+        "device_class": "current",
+        "state_class": "measurement",
+        "icon": "mdi:battery-charging",
+    },
+    "Battery Discharge Current": {
+        "name": "Battery Discharge Current",
+        "unit": UNITS["CURRENT"],
+        "device_class": "current",
+        "state_class": "measurement",
+        "icon": "mdi:battery-arrow-down",
+    },
     "Batt Current": {
         "name": "Battery Current",
         "unit": UNITS["CURRENT"],
@@ -198,6 +212,13 @@ SENSOR_TYPES = {
         "state_class": "measurement",
         "icon": "mdi:sine-wave",
     },
+    "Output Frequency": {
+        "name": "Output Frequency",
+        "unit": UNITS["FREQUENCY"],
+        "device_class": "frequency",
+        "state_class": "measurement",
+        "icon": "mdi:sine-wave",
+    },
     "Inverter frequency": {
         "name": "Inverter Frequency",
         "unit": UNITS["FREQUENCY"],
@@ -226,6 +247,12 @@ SENSOR_TYPES = {
         "icon": "mdi:gauge",
     },
     "SOC": {
+        "name": "State of Charge",
+        "unit": UNITS["PERCENTAGE"],
+        "state_class": "measurement",
+        "icon": "mdi:battery-high",
+    },
+    "State of Charge": {
         "name": "State of Charge",
         "unit": UNITS["PERCENTAGE"],
         "state_class": "measurement",
@@ -457,6 +484,20 @@ SENSOR_TYPES = {
         "device_class": "energy",
         "state_class": "total_increasing",
         "icon": "mdi:flash",
+    },
+    "Energy Today": {
+        "name": "Energy Today",
+        "unit": UNITS["ENERGY"],
+        "device_class": "energy",
+        "state_class": "total_increasing",
+        "icon": "mdi:solar-power",
+    },
+    "Energy Total": {
+        "name": "Energy Total",
+        "unit": UNITS["ENERGY"],
+        "device_class": "energy",
+        "state_class": "total_increasing",
+        "icon": "mdi:solar-power",
     },
     "Battery Energy Today (Charge)": {
         "name": "Battery Energy Today (Charge)",